Reorganized the Close Menu, Quit, and Reset buttons to match the layout in SoH. (#1344)
* Reorganized the Menu "Reset","Quit 2S2H , and "Closed Menu" buttons
Place in the same order as it is in SoH "Quit, Reset, and Close Menu" on
the right.

* Ported "SohModals" from SoH and Renamed it to "BenModals

* When the "Quit" Button is pressed it will pop up a message asking the
Player "Are you sure you want to quit 2S2H?". with two buttons
"Quit" and "Cancel" like how it is done in SoH.

* Cleaned up unneeded "includes" in "BenModals.h" and "BenModals.cpp"
Removed "extern "C" PlayState* gPlayState;" from "BenModals.cpp"
it wasn't used for anything.

---------

Co-authored-by: Eblo <7004497+Eblo@users.noreply.github.com>